pyproject.toml
[tool.poetry]
name = "payment-service"
version = "2023.1"
description = ""
authors = ["oleg.korshunov"]

[tool.poetry.dependencies]
python = "~3.11.1"
fastapi = "^0.94.0"
orjson = "^3.8.6"
uvicorn = {extras = ["standard"], version = "^0.21.1"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.9"}
alembic = "^1.10.2"
asyncpg = "^0.27.0"
prometheus-client = "^0.16.0"
prometheus-fastapi-instrumentator = "^6.0.0"
aiokafka = "^0.8.1"
pika = "^1.3.2"
pre-commit = "^3.5.0"
redis = "^5.1.0"

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
isort = "^5.12.0"
mypy = "^1.0.1"
pylint = "^2.16.2"
pytest = "^7.2.1"
pytest-env = "^0.8.1"
pytest-cov = "^4.0.0"
pytest-asyncio = "^0.20.3"
httpx = "^0.23.3"
pytest-factoryboy = "^2.5.1"
types-redis = "^4.6.0.20240903"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ["py311"]
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
)
'''

[tool.isort]
line_length = 120
profile = "black"
skip_glob = ["migrations/versions/*"]

[tool.mypy]
ignore_missing_imports = true
exclude = [
    'venv/*',
    'tests/*',
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = "--cov app --cov tests --cov-report term-missing --cov-branch --junitxml=test-reports/junit.xml"
env = [
  "APP_DEBUG = 0",
  "APP_JSON_LOGGER_ENABLED = 0",
  "DOC_STORAGE_HOST="
]

[tool.coverage.report]
show_missing = true
exclude_lines = [
  "pragma: no cover",
  "if TYPE_CHECKING:",
  "@abc.abstractmethod",
  'if __name__ == "__main__":',
]

[tool.coverage.run]
concurrency = [
  "thread",
  "greenlet",
]
